***PRY FLEXIBILITY SRL*** -VENTAS CONTROLLER

+++ FlexibilitySRL/src/main/java/com/flexibilitysrl/models/Clientes.java
@@ -5,12 +5,13 @@ import lombok.Getter;
 import lombok.NoArgsConstructor;
 import lombok.Setter;
 
+
 @Getter
 @Setter
 @AllArgsConstructor
 @NoArgsConstructor
 public class Clientes {
-    private int idCliente;
+    private Long idCliente;
     private Persona persona;
     private String razonSocial;
 }
